app/requirements.txt
-e .
aws_cdk.aws_s3
aws_cdk.aws_iam
